viewModelを追加し、コードを整理
